refactor: enhance compression logic, test coverage, and config consistency
- Replaced `for` loop with `while` in `BitwiseOps` for clearer flow.
- Adjusted Detekt rules (activated unused code checks, zero max issues).
- Improved test assertions in `BitUtilsImprovedTest` and `MaxWBitsImportTest`.
- Streamlined string interpolation in `ZLibCli`.
- Updated `build.gradle.kts` to enforce checks on all source sets, including tests.
- Refined doc formatting and suppressed unused/empty checks.

+++ src/commonMain/kotlin/ai/solace/zlib/bitwise/checksum/Adler32Utils.kt
@@ -36,7 +36,7 @@ object Adler32Utils {
         ): Long {
             if (buf == null) return 1L
 
-            var a = (adler and 0xFFFF).toLong()
+            var a = (adler and 0xFFFF)
             var b = ((adler ushr 16) and 0xFFFF)
 
             var i = index
